ensure that the importage being modified is owned by the user modifying it

+++ app/controllers/importages_controller.rb
@@ -1,5 +1,6 @@
 class ImportagesController < ApplicationController
   before_action :set_importage, only: [:show, :edit, :update, :destroy]
+  before_action :authenticate_musician!, :except => [:show]
 
   # GET /importages
   # GET /importages.json
@@ -27,7 +28,7 @@ class ImportagesController < ApplicationController
     @importage = Importage.new(importage_params)
 
     respond_to do |format|
-      if @importage.save
+      if ownership_check(@importage) && @importage.save
         format.html { redirect_to @importage, notice: 'Importage was successfully created.' }
         format.json { render action: 'show', status: :created, location: @importage }
       else
@@ -40,8 +41,9 @@ class ImportagesController < ApplicationController
   # PATCH/PUT /importages/1
   # PATCH/PUT /importages/1.json
   def update
+
     respond_to do |format|
-      if @importage.update(importage_params)
+      if ownership_check(@importage) && @importage.update(importage_params)
         format.html { redirect_to @importage, notice: 'Importage was successfully updated.' }
         format.json { head :no_content }
       else
@@ -54,21 +56,43 @@ class ImportagesController < ApplicationController
   # DELETE /importages/1
   # DELETE /importages/1.json
   def destroy
-    eximportage = @importage.destroy
-    respond_to do |format|
-      format.html { redirect_to importages_url }
-      format.json { render :text => "{\"deleted\": #{eximportage.id}}" }
+    if ownership_check(@importage)
+      eximportage = @importage.destroy
+      respond_to do |format|
+        format.html { redirect_to importages_url }
+        format.json { render :text => "{\"deleted\": #{eximportage.id}}" }
+      end
+    else
+      respond_to do |format|
+        format.json { render :text => "{\"not-deleted\": #{@importage.id}}" }
+      end
     end
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_importage
-      @importage = Importage.find(params[:id])
-    end
+  # Use callbacks to share common setup or constraints between actions.
+  def set_importage
+    @importage = Importage.find(params[:id])
+  end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def importage_params
-      params.require(:importage).permit(:start_time, :simple_volume)
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def importage_params
+    params.require(:importage).permit(:start_time, :simple_volume)
+  end
+
+  # make sure the current_musician has the rights to edit this
+  def ownership_check(importage)
+     unless importage.library_id.nil? ^ importage.mix_id.nil?
+      raise 'importage #{importage.id} is orphaned or binding to a sample to a library and a mix'
     end
+    # if a mix importage verify mix.musican
+    if importage.mix_id && (importage.mix.musician_id == current_musician.id)
+      true
+      # if a library importage verify library.musican
+    elsif importage.library_id && (importage.library.musician_id == current_musician.id)
+      true
+    else
+      false
+    end
+  end
 end
